BACKEND/requirements.txt
 # Руководство к файлу (BACKEND/requirements.txt)
 # Назначение:
 # - Список Python-зависимостей backend VKMax для установки в Docker и локальной разработке.
 # - Не должен содержать локальных editable-путей (формата `-e /home/...`), чтобы сборка Docker не падала.
 # - Все дополнительные dev-зависимости можно хранить отдельно (например, в requirements-dev.txt).

 aiohappyeyeballs==2.6.1
 aiohttp==3.13.2
 aiolimiter==1.2.1
 aiomax
 asyncpg
 aiosignal==1.4.0
 aiosqlite==0.21.0
 annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
cobble==0.1.4
fastapi==0.121.2
filelock==3.20.0
fire==0.7.1
fonttools==4.60.1
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
lxml==6.0.2
mammoth==1.11.0
multidict==6.7.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
orjson==3.11.4
packaging==25.0
pdf2docx==0.5.8
pdfkit==1.0.0
pillow==12.0.0
playwright==1.56.0
pluggy==1.6.0
propcache==0.4.1
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
PyMuPDF==1.26.6
pytest==9.0.1
pytest-asyncio==1.3.0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
reportlab==4.4.4
requests==2.32.5
requests-file==3.0.1
selectolax==0.4.1
 sniffio==1.3.1
 soupsieve==2.8
 SQLAlchemy==2.0.44
 starlette==0.49.3
 termcolor==3.2.0
 tldextract==5.3.0
 typing-inspection==0.4.2
 typing_extensions==4.15.0
 urllib3==2.5.0
 uvicorn==0.38.0
 uvloop==0.22.1
 watchfiles==1.1.1
 # Editable install with no version control (webparser==0.1.0)
 # -e /home/alexander/Projects/WebParser  # локальная dev-зависимость; отключена для Docker-сборки
 websockets==15.0.1
 wkhtmltopdf==0.2
 yarl==1.22.0
